[#84] Updated logic for using extended sources
- added configuration for new source_types
